+++ src/main/java/net/canadensys/dataportal/vascan/dao/NameDAO.java
@@ -14,21 +14,36 @@ import net.canadensys.query.LimitedResult;
 public interface NameDAO {
 
 	/**
-	 * Search NameModel from a String
+	 * Search NameModel from a String.
 	 * @param text
+	 * @param useAutocompletion should we use AutoCompletion feature to return partially matched names. 
 	 * @return
 	 */
-	public LimitedResult<List<NameConceptModelIF>> search(String text);
+	public LimitedResult<List<NameConceptModelIF>> search(String text, boolean useAutocompletion);
+	
 	/**
 	 * Search NameModel from a String with paging.
 	 * @param text
+	 * @param useAutocompletion should we use AutoCompletion feature to return partially matched names. 
 	 * @param pageNumber (starting at 0)
 	 * @return
 	 */
-	public LimitedResult<List<NameConceptModelIF>> search(String text, int pageNumber);
+	public LimitedResult<List<NameConceptModelIF>> search(String text, boolean useAutocompletion, int pageNumber);
 	
-	
+	/**
+	 * Search only taxon type.
+	 * Autocompletion will be used.
+	 * @param text
+	 * @return
+	 */
 	public List<NameConceptModelIF> searchTaxon(String text);
+	
+	/**
+	 * Search only vernacular type.
+	 * Autocompletion will be used.
+	 * @param text
+	 * @return
+	 */
 	public List<NameConceptModelIF> searchVernacular(String text);
 	
 	/**