+++ HISTORY.md
@@ -1,5 +1,8 @@
 Version History
 ===============
+Version 2.1.0 (not yet released)
+* ElasticSearchNameDAO search function now includes epithet and "genus first letter"
+* Search functions in NameDAO now allow to include or not ngram index.
 
 Version 2.0.1 2013-08-23
 * Add support for synonym with more than one parent in NameDAO and model